requirements.txt
PyQt6>=6.7
PyYAML>=6.0
